fixup! Support org.w3c.dom.Document input to SVGImageReader
Extract SVGImageReaderSpi DOM tests into separate class/suite.
