+++ imageio/imageio-batik/src/test/java/com/twelvemonkeys/imageio/plugins/svg/SVGImageReaderSpiTest.java
@@ -34,10 +34,6 @@ import com.twelvemonkeys.imageio.stream.ByteArrayImageInputStream;
 import com.twelvemonkeys.imageio.stream.URLImageInputStreamSpi;
 import org.junit.Test;
 
-import org.w3c.dom.Document;
-import javax.xml.parsers.DocumentBuilder;
-import javax.xml.parsers.DocumentBuilderFactory;
-
 import javax.imageio.ImageIO;
 import javax.imageio.spi.IIORegistry;
 import javax.imageio.spi.ImageReaderSpi;
@@ -56,7 +52,7 @@ import static org.junit.Assert.assertTrue;
  */
 public class SVGImageReaderSpiTest {
 
-    private static final String[] VALID_INPUTS = {
+    static final String[] VALID_INPUTS = {
             "/svg/Android_robot.svg", // Minimal, no xml dec, no namespace
             "/svg/batikLogo.svg",     // xml dec, comments, namespace
             "/svg/blue-square.svg",   // xml dec, namespace
@@ -100,59 +96,5 @@ public class SVGImageReaderSpiTest {
                 assertFalse("Claims to read invalid input:" + invalidInput, provider.canDecodeInput(input));
             }
         }
-    }
-
-    @Test
-    public void canDecodeSVGDocument() throws Exception {
-        DocumentBuilderFactory dbf = DocumentBuilderFactory.newDefaultInstance();
-        dbf.setNamespaceAware(true);
-
-        DocumentBuilder domBuilder = dbf.newDocumentBuilder();
-        for (String validInput : VALID_INPUTS) {
-            Document input = domBuilder.parse(getClass().getResource(validInput).toString());
-            assertTrue("Can't read valid input: " + validInput, provider.canDecodeInput(input));
-        }
-    }
-
-    @Test
-    public void cannotDecodeNamespaceUnawareDocument() throws Exception {
-        DocumentBuilderFactory dbf = DocumentBuilderFactory.newDefaultInstance();
-
-        DocumentBuilder domBuilder = dbf.newDocumentBuilder();
-        for (String validInput : VALID_INPUTS) {
-            Document input = domBuilder.parse(getClass().getResource(validInput).toString());
-            assertFalse("Claims to read namespace unaware document: " + validInput,
-                        provider.canDecodeInput(input));
-        }
-    }
-
-    @Test
-    public void cannotDecodeNonSVGNamespaceDocument() throws Exception {
-        DocumentBuilderFactory dbf = DocumentBuilderFactory.newDefaultInstance();
-        dbf.setNamespaceAware(true);
-
-        Document document = dbf.newDocumentBuilder().newDocument();
-        document.appendChild(document
-                .createElementNS("http://www.w3.org/2000/doovde", "svg"));
-
-        assertFalse("Claims to read document with root element: {"
-                + document.getDocumentElement().getNamespaceURI() + "}"
-                + document.getDocumentElement().getLocalName(),
-                provider.canDecodeInput(document));
-    }
-
-    @Test
-    public void cannotDecodeNonSVGRootElement() throws Exception {
-        DocumentBuilderFactory dbf = DocumentBuilderFactory.newDefaultInstance();
-        dbf.setNamespaceAware(true);
-
-        Document document = dbf.newDocumentBuilder().newDocument();
-        document.appendChild(document
-                .createElementNS("http://www.w3.org/2000/svg", "path"));
-
-        assertFalse("Claims to read document with root element: {"
-                + document.getDocumentElement().getNamespaceURI() + "}"
-                + document.getDocumentElement().getLocalName(),
-                provider.canDecodeInput(document));
     }
 }
